Moving the dashboard-related components out of the 'widgets' folder and into a 'dashboard' components folder.
